Migrate build tooling from CRA/react-scripts to Vite
Resolves the remaining 25 Dependabot alerts, which were all transitive
dev-only dependencies of react-scripts (webpack-dev-server, node-forge,
minimatch, svgo, lodash, etc.) with no fix available short of breaking
the build.

- Replace react-scripts with vite + vitest, drop web-vitals/reportWebVitals
- Move index.html to project root, rename JSX-containing .js files to .jsx
- Rename REACT_APP_* env vars to VITE_* and use import.meta.env
- Update npm scripts (start/build/test/preview)
