+++ .env.example
@@ -1,5 +1,5 @@
-REACT_APP_MAILCHIMP_URL="https://gmail.us10.list-manage.com/subscribe/post"
-REACT_APP_MAILCHIMP_U="u"
-REACT_APP_MAILCHIMP_ID="your_mailchimp_id"
+VITE_MAILCHIMP_URL="https://gmail.us10.list-manage.com/subscribe/post"
+VITE_MAILCHIMP_U="u"
+VITE_MAILCHIMP_ID="your_mailchimp_id"
 
-REACT_APP_WEB3FORMS_ACCESS_KEY=your_web3forms_access_key
+VITE_WEB3FORMS_ACCESS_KEY=your_web3forms_access_key
