+++ src/App.test.jsx
@@ -1,0 +1,6 @@
+import { render } from '@testing-library/react';
+import App from './App';
+
+test('renders app without crashing', () => {
+  render(<App />);
+});
