+++ src/components/Banner.jsx
